+++ @utahdts/utah-design-system/react/components/table/TableFilterDate.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { joinClassNames } from '../../util/joinClassNames';
 import { TextInput } from '../forms/TextInput';
+import { useTableContext } from './hooks/useTableContext';
 import { useCurrentValuesFromStateContext } from './useCurrentValuesFromStateContext';
 
 /**
@@ -10,6 +11,7 @@ import { useCurrentValuesFromStateContext } from './useCurrentValuesFromStateCon
  * @param {TableDataT} [props.defaultValue]
  * @param {import('react').RefObject<HTMLTableCellElement>} [props.innerRef]
  * @param {string} [props.id]
+ * @param {string} props.a11yLabel This should be an accessibility readable field name. 'Filter' will be prepended to it.
  * @param {(e: React.ChangeEvent) => TableDataT} [props.onChange]
  * @param {string} props.recordFieldPath
  * @param {TableDataT} [props.value]
@@ -20,6 +22,7 @@ export function TableFilterDate({
   defaultValue,
   innerRef,
   id,
+  a11yLabel,
   onChange,
   recordFieldPath,
   value,
@@ -36,11 +39,12 @@ export function TableFilterDate({
     onChange,
     value,
   });
+  const { state: { tableId } } = useTableContext();
   return (
     <th className={joinClassNames('table-header__cell table-header__cell--filter-date', className)} id={id ?? undefined} ref={innerRef}>
       <TextInput
-        id={`table-filter-date-${recordFieldPath}`}
-        label={`filter ${recordFieldPath}`}
+        id={`${tableId}__table-filter-date-${recordFieldPath}`}
+        label={`Filter ${a11yLabel}`}
         onChange={currentOnChange}
         value={currentValue?.toString()}
         // eslint-disable-next-line react/jsx-props-no-spreading