+++ @utahdts/utah-design-system/react/components/table/TableFilterTextInput.jsx
@@ -12,6 +12,7 @@ import { useTableContext } from './hooks/useTableContext';
  * @param {boolean} [props.exactMatch]
  * @param {import('react').RefObject<HTMLTableCellElement>} [props.innerRef]
  * @param {string} [props.id]
+ * @param {string} props.a11yLabel This should be an accessibility readable field name. 'Filter' will be prepended to it.
  * @param {((e: React.ChangeEvent) => (string | void | undefined))} [props.onChange]
  * @param {string} props.recordFieldPath
  * @param {string} [props.value]
@@ -23,6 +24,7 @@ export function TableFilterTextInput({
   exactMatch,
   innerRef,
   id,
+  a11yLabel,
   onChange,
   recordFieldPath,
   value,
@@ -40,15 +42,14 @@ export function TableFilterTextInput({
     onChange,
     value,
   });
-  const { state: { tableWrapperId } } = useTableContext();
-
   useTableFilterRegistration(recordFieldPath, !!exactMatch, defaultValue);
+  const { state: { tableId } } = useTableContext();
 
   return (
     <th className={joinClassNames('table-header__cell table-header__cell--filter-text-input', className)} id={id ?? undefined} ref={innerRef}>
       <TextInput
-        id={`${tableWrapperId}__table-filter-text-input-${recordFieldPath}`}
-        label={`filter ${recordFieldPath}`}
+        id={`${tableId}__table-filter-text-input-${recordFieldPath}`}
+        label={`Filter ${a11yLabel || recordFieldPath}`}
         onChange={currentOnChange}
         onClear={() => setValue('')}
         value={currentValue}